lucy/app/requirements.txt
python-dotenv
pinecone
google-generativeai
requests
Flask
# Optional: Gunicorn for production deployment with Flask, Vercel might handle this
# gunicorn
